VIRO-2350: Legacy Adreno support for stereoscopic
Support stereoscopic video in legacy Adreno
devices. The video had a purple tint in stereoscopic
mode. Two changes made it go away:

1. Got rid of the 'if' statement in stereoscopic and
    replaced with a simple multiplication. Branching in
    shaders is generally inefficient anyway so this is
    an overall improvement even for higher end devices.

2. Made the eye_type a floating point *highp* value.
    For some reason lowp causes a purple tint whereas
    highp makes it go away. This is an internal driver
    bug.

+++ ViroRenderer/VROShaderProgram.cpp
@@ -509,7 +509,7 @@ void VROShaderProgram::addStandardUniforms() {
     addUniform(VROShaderProperty::Mat4, 1, "view_matrix");
     addUniform(VROShaderProperty::Mat4, 1, "projection_matrix");
     addUniform(VROShaderProperty::Vec3, 1, "camera_position");
-    addUniform(VROShaderProperty::Int, 1, "eye_type");
+    addUniform(VROShaderProperty::Float, 1, "eye_type");
 
     addUniform(VROShaderProperty::Vec4, 1, "material_diffuse_surface_color");
     addUniform(VROShaderProperty::Float, 1, "material_diffuse_intensity");
